+++ HW8/treeAVL/dictionary.h
@@ -1,0 +1,6 @@
+#ifndef DICTIONARY_H
+#define DICTIONARY_H
+
+void dictionary(void);
+
+#endif
