+++ HW8/treeAVL/tests.h
@@ -1,0 +1,7 @@
+#ifndef TESTS_H
+#define TESTS_H
+#include <stdbool.h>
+
+bool runTests(void);
+
+#endif //TESTS_H
