+++ HW8/treeAVL2/dictionary.h
@@ -1,0 +1,8 @@
+//
+// Created by Михаил Харламов on 05.12.2024.
+//
+
+#ifndef DICTIONARY_H
+#define DICTIONARY_H
+
+#endif //DICTIONARY_H
