+++ HW8/treeAVL2/tests.h
@@ -1,0 +1,8 @@
+//
+// Created by Михаил Харламов on 05.12.2024.
+//
+
+#ifndef TESTS_H
+#define TESTS_H
+
+#endif //TESTS_H
